Use videoView.start instead of videoView.resume.
Currently when the screen turns off, we are triggering videoView.stopPlayback
which will release the mediaplayer instance. And when the screen turns
on, we trigger videoView.resume which will create a new mediaplayer
instance. This seems quite unnecessary as we are going to turn on the screen
immediately if the user turns off the screen.

Bug: 29359781
Change-Id: I63db916f11c0f3c6d07d6d18c7d8e567ed8b4318